Added details for loadIds
